BackEnd/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt


